fix: auth client secrets as keel secrets
